pyproject.toml
[project]
name = "document-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "赵亚同",email = "yatong.zhao@foxmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "llama-index-core (>=0.12.43,<0.13.0)",
    "llama-index-llms-deepseek (>=0.1.2,<0.2.0)",
    "llama-index-utils-workflow (>=0.3.4,<0.4.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "chromadb (>=1.0.13,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "markdown (>=3.8.2,<4.0.0)",
    "llama-index-vector-stores-chroma (>=0.4.2,<0.5.0)",
    "llama-index-embeddings-dashscope (>=0.3.0,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
